Rename trigger -> track, unbundle
